Melody/driving_test_refresh/requirements.txt
playwright>=1.49.1
httpx>=0.28.1
python-dotenv>=1.0.1
